Makefile
This works with rendering a white window in a standalone app. I am going to be adding more and more things slowly and make sure they render.

The reason why I am testing things like this is that since I have so many issues withe generating a standalone app, a project like this helps